+++ Engine/src/sections/drawEmitters.cpp
@@ -104,7 +104,7 @@ namespace Phoenix {
 
 		m_pExprPosition = new MathDriver(this);
 		// Load all the other strings
-		for (int i = 2; i < strings.size(); i++)
+		for (size_t i = 2; i < strings.size(); i++)
 			m_pExprPosition->expression += strings[i];
 
 		m_pExprPosition->SymbolTable.add_variable("tx", m_vTranslation.x);
